feat: add Prompt Templates and Skills tabs to FAC Admin page (#135)
* feat: add Prompt Templates and Skills tabs to FAC Admin page

- Add top-level tabs (Tools, Prompt Templates, Skills) to the admin page
- Prompt Templates tab: list with search/filter, publish/unpublish toggle,
  inline markdown preview with arguments display, link to DocType form
- Skills tab: list with type/status filters, publish/unpublish toggle,
  inline markdown content viewer with proper table rendering, link to DocType
- Add 5 new admin API endpoints with role guards and POST restrictions:
  get_prompt_templates_list, toggle_prompt_template_status,
  preview_prompt_template, get_skills_list, toggle_skill_status
- Add role guard to get_server_settings for consistency
- Stats grid: 5 cards in a row with responsive breakpoints
- Clean card layout with visual hierarchy (title, subtitle, metadata chips)
- Custom Showdown converter with tables enabled for proper markdown rendering
- Full dark mode support with explicit overrides for all new components

* refactor: split admin_api.py and fac_admin.js into modular files

Python: Split admin_api.py (952 lines) into 6 submodules under
api/admin/ (server, plugins, stats, tools, prompts, skills).
admin_api.py is now a thin re-export facade preserving all
frappe.call method paths and test imports.

JavaScript: Split fac_admin.js (2398 lines) into an orchestrator
(388 lines) + 4 submodules loaded via frappe.require():
- public/js/fac_admin_utils.js (renderMarkdown helper)
- public/js/fac_admin_tools.js (tools/plugins tab, server status)
- public/js/fac_admin_prompts.js (prompt templates tab)
- public/js/fac_admin_skills.js (skills tab)

CSS: Extracted 815 lines to public/css/fac_admin.css loaded via
frappe.require. Submodules communicate via frappe.fac_admin namespace.

* chore: ignore node_modules symlinks in public/

* fix: add type hints to all whitelisted admin API functions

Adds parameter and return type annotations to satisfy the Frappe
semgrep rule missing-argument-type-hint on all whitelisted endpoints.

* ci: trigger re-run with corrected PR base branch

---------
