+++ .gitignore
@@ -34,6 +34,7 @@ venv.bak/
 
 # Dependency directories
 node_modules/
+node_modules
 jspm_packages/
 
 # IDEs and editors
