+++ frappe_assistant_core/api/admin/tools.py
@@ -1,0 +1,502 @@
+# Frappe Assistant Core - AI Assistant integration for Frappe Framework
+# Copyright (C) 2025 Paul Clinton
+#
+# AGPL-3.0-or-later — see <https://www.gnu.org/licenses/>.
+
+import frappe
+from frappe import _
+
+
+@frappe.whitelist()
+def get_tool_configurations() -> dict:
+    """
+    Get all tool configurations with detailed information.
+
+    Returns a list of tools with their configuration status, category,
+    and role access settings.
+    """
+    frappe.only_for(["System Manager", "Assistant Admin"])
+    from frappe_assistant_core.core.tool_registry import get_tool_registry
+    from frappe_assistant_core.utils.plugin_manager import get_plugin_manager
+    from frappe_assistant_core.utils.tool_category_detector import get_category_info
+
+    try:
+        plugin_manager = get_plugin_manager()
+        tool_registry = get_tool_registry()
+        all_tools = plugin_manager.get_all_tools()
+        enabled_plugins = plugin_manager.get_enabled_plugins()
+
+        # Include external tools from hooks (registered via assistant_tools)
+        external_tools = tool_registry._get_external_tools()
+        all_tools.update(external_tools)
+
+        # Get all existing configurations
+        existing_configs = {}
+        if frappe.db.table_exists("FAC Tool Configuration"):
+            configs = frappe.get_all(
+                "FAC Tool Configuration",
+                fields=[
+                    "name",
+                    "tool_name",
+                    "plugin_name",
+                    "enabled",
+                    "tool_category",
+                    "auto_detected_category",
+                    "category_override",
+                    "role_access_mode",
+                    "description",
+                ],
+            )
+            for config in configs:
+                tool_name = config.get("tool_name") or config.get("name")
+                existing_configs[tool_name] = config
+
+        tool_list = []
+        for tool_name, tool_info in all_tools.items():
+            plugin_enabled = tool_info.plugin_name in enabled_plugins
+
+            # Get configuration if exists
+            config = existing_configs.get(tool_name, {})
+            tool_enabled = config.get("enabled", 1) if config else 1
+            category = config.get("tool_category", "read_write") if config else "read_write"
+            # Normalize 'dangerous' to 'privileged' for UI consistency
+            if category == "dangerous":
+                category = "privileged"
+            auto_category = config.get("auto_detected_category", "") if config else ""
+            if auto_category == "dangerous":
+                auto_category = "privileged"
+            category_override = config.get("category_override", 0) if config else 0
+            role_access_mode = config.get("role_access_mode", "Allow All") if config else "Allow All"
+
+            # Get role access if configured
+            role_access = []
+            if config:
+                try:
+                    role_access = frappe.get_all(
+                        "FAC Tool Role Access",
+                        filters={"parent": config.get("name")},
+                        fields=["role", "allow_access"],
+                    )
+                except Exception:
+                    pass
+
+            # Get category display info
+            category_info = get_category_info(category)
+
+            tool_list.append(
+                {
+                    "name": tool_name,
+                    "display_name": tool_name.replace("_", " ").title(),
+                    "description": tool_info.description,
+                    "plugin_name": tool_info.plugin_name,
+                    "plugin_display_name": tool_info.plugin_name.replace("_", " ").title(),
+                    "plugin_enabled": plugin_enabled,
+                    "tool_enabled": bool(tool_enabled),
+                    "effectively_enabled": plugin_enabled and bool(tool_enabled),
+                    "category": category,
+                    "category_label": category_info.get("label", "Unknown"),
+                    "category_color": category_info.get("color", "gray"),
+                    "category_icon": category_info.get("icon", "fa-question"),
+                    "auto_detected_category": auto_category,
+                    "category_override": bool(category_override),
+                    "role_access_mode": role_access_mode,
+                    "role_access": role_access,
+                    "has_config": bool(config),
+                }
+            )
+
+        # Sort by plugin name, then by tool name
+        tool_list.sort(key=lambda x: (x["plugin_name"], x["name"]))
+
+        return {"success": True, "tools": tool_list}
+
+    except Exception as e:
+        frappe.log_error(f"Failed to get tool configurations: {str(e)}")
+        return {"success": False, "error": str(e), "tools": []}
+
+
+@frappe.whitelist(methods=["POST"])
+def toggle_tool(tool_name: str, enabled: bool):
+    """
+    Enable or disable an individual tool.
+
+    Args:
+        tool_name: The name of the tool to toggle
+        enabled: True to enable, False to disable
+
+    Returns:
+        Success status and message
+    """
+    frappe.only_for(["System Manager", "Assistant Admin"])
+    from frappe_assistant_core.core.tool_registry import get_tool_registry
+    from frappe_assistant_core.utils.plugin_manager import get_plugin_manager
+    from frappe_assistant_core.utils.tool_category_detector import detect_tool_category
+
+    try:
+        # Validate tool exists
+        plugin_manager = get_plugin_manager()
+        tool_registry = get_tool_registry()
+        all_tools = plugin_manager.get_all_tools()
+
+        # Include external tools from hooks
+        external_tools = tool_registry._get_external_tools()
+        all_tools.update(external_tools)
+
+        if tool_name not in all_tools:
+            return {"success": False, "message": _(f"Tool '{tool_name}' not found")}
+
+        # Convert enabled to boolean
+        enabled = frappe.utils.cint(enabled)
+
+        # Use savepoint for atomic operation
+        frappe.db.savepoint("toggle_tool")
+
+        try:
+            # Get or create tool configuration
+            if frappe.db.exists("FAC Tool Configuration", tool_name):
+                config = frappe.get_doc("FAC Tool Configuration", tool_name)
+                config.enabled = enabled
+                config.save(ignore_permissions=True)
+            else:
+                # Create new configuration
+                tool_info = all_tools[tool_name]
+                category = detect_tool_category(tool_info.instance)
+
+                config = frappe.new_doc("FAC Tool Configuration")
+                config.tool_name = tool_name
+                config.plugin_name = tool_info.plugin_name
+                config.description = tool_info.description
+                config.enabled = enabled
+                config.tool_category = category
+                config.auto_detected_category = category
+                config.source_app = getattr(tool_info.instance, "source_app", "frappe_assistant_core")
+                config.insert(ignore_permissions=True)
+
+            frappe.db.release_savepoint("toggle_tool")
+            frappe.db.commit()
+
+            # Clear caches
+            tool_registry = get_tool_registry()
+            tool_registry.clear_cache()
+
+            cache = frappe.cache()
+            cache.delete_keys("fac_tool_*")
+
+            action = "enabled" if enabled else "disabled"
+            return {"success": True, "message": _(f"Tool '{tool_name}' {action} successfully")}
+
+        except Exception as e:
+            frappe.db.rollback_savepoint("toggle_tool")
+            raise e
+
+    except Exception as e:
+        frappe.log_error(f"Failed to toggle tool '{tool_name}': {str(e)}")
+        return {"success": False, "message": _(f"Error: {str(e)}")}
+
+
+@frappe.whitelist(methods=["POST"])
+def bulk_toggle_tools(tool_names: list, enabled: bool):
+    """
+    Enable or disable multiple tools at once.
+
+    Args:
+        tool_names: List of tool names to toggle
+        enabled: True to enable, False to disable
+
+    Returns:
+        Success status with details
+    """
+    frappe.only_for(["System Manager", "Assistant Admin"])
+    if isinstance(tool_names, str):
+        import json
+
+        tool_names = json.loads(tool_names)
+
+    results = {"success": True, "toggled": [], "failed": []}
+
+    for tool_name in tool_names:
+        result = toggle_tool(tool_name, enabled)
+        if result.get("success"):
+            results["toggled"].append(tool_name)
+        else:
+            results["failed"].append({"name": tool_name, "error": result.get("message")})
+
+    if results["failed"]:
+        results["success"] = False
+        results["message"] = _(f"Failed to toggle {len(results['failed'])} tools")
+    else:
+        action = "enabled" if enabled else "disabled"
+        results["message"] = _(f"Successfully {action} {len(results['toggled'])} tools")
+
+    return results
+
+
+@frappe.whitelist(methods=["POST"])
+def bulk_toggle_tools_by_category(category: str = None, enabled: bool = True, plugin_name: str = None):
+    """
+    Enable or disable all tools in a category.
+
+    Args:
+        category: Tool category (read_only, write, read_write, privileged) or None for all
+        enabled: True to enable, False to disable
+        plugin_name: Optional - filter by plugin (e.g., "core", "data_science")
+
+    Returns:
+        {"success": bool, "toggled": [...], "failed": [...], "total": int, "message": str}
+    """
+    frappe.only_for(["System Manager", "Assistant Admin"])
+
+    # Parse JSON if passed as string
+    if isinstance(enabled, str):
+        enabled = enabled.lower() in ("true", "1", "yes")
+
+    # Validate category if provided
+    valid_categories = ["read_only", "write", "read_write", "privileged"]
+    if category and category not in valid_categories:
+        return {
+            "success": False,
+            "message": _(f"Invalid category. Must be one of: {', '.join(valid_categories)}"),
+        }
+
+    # Build filters for query
+    filters = {}
+    if category:
+        filters["tool_category"] = category
+    if plugin_name:
+        filters["plugin_name"] = plugin_name
+
+    # Get matching tools from FAC Tool Configuration
+    try:
+        if filters:
+            tool_names = frappe.get_all("FAC Tool Configuration", filters=filters, pluck="tool_name")
+        else:
+            tool_names = frappe.get_all("FAC Tool Configuration", pluck="tool_name")
+    except Exception as e:
+        return {"success": False, "message": _(f"Error querying tools: {str(e)}")}
+
+    if not tool_names:
+        filter_desc = []
+        if category:
+            filter_desc.append(f"category '{category}'")
+        if plugin_name:
+            filter_desc.append(f"plugin '{plugin_name}'")
+        filter_str = " and ".join(filter_desc) if filter_desc else "the given criteria"
+        return {
+            "success": True,
+            "toggled": [],
+            "failed": [],
+            "total": 0,
+            "message": _(f"No tools found matching {filter_str}"),
+        }
+
+    # Toggle each tool
+    results = {"success": True, "toggled": [], "failed": [], "total": len(tool_names)}
+
+    for tool_name in tool_names:
+        result = toggle_tool(tool_name, enabled)
+        if result.get("success"):
+            results["toggled"].append(tool_name)
+        else:
+            results["failed"].append({"name": tool_name, "error": result.get("message")})
+
+    # Set overall status and message
+    if results["failed"]:
+        results["success"] = len(results["toggled"]) > 0  # Partial success
+        results["message"] = _(f"{len(results['toggled'])} tools toggled, {len(results['failed'])} failed")
+    else:
+        action = "enabled" if enabled else "disabled"
+        filter_desc = []
+        if category:
+            filter_desc.append(f"'{category}'")
+        if plugin_name:
+            filter_desc.append(f"in '{plugin_name}'")
+        filter_str = " ".join(filter_desc) if filter_desc else ""
+        results["message"] = _(f"Successfully {action} {len(results['toggled'])} {filter_str} tools")
+
+    return results
+
+
+@frappe.whitelist(methods=["POST"])
+def update_tool_category(tool_name: str, category: str, override: bool = True):
+    """
+    Update the category for a tool.
+
+    Args:
+        tool_name: The name of the tool
+        category: The new category (read_only, write, read_write, privileged)
+        override: Whether to mark this as a manual override
+
+    Returns:
+        Success status and message
+    """
+    frappe.only_for(["System Manager", "Assistant Admin"])
+    from frappe_assistant_core.core.tool_registry import get_tool_registry
+    from frappe_assistant_core.utils.plugin_manager import get_plugin_manager
+    from frappe_assistant_core.utils.tool_category_detector import detect_tool_category
+
+    # Accept both 'privileged' and 'dangerous' for backward compatibility
+    valid_categories = ["read_only", "write", "read_write", "privileged", "dangerous"]
+    # Normalize 'dangerous' to 'privileged'
+    if category == "dangerous":
+        category = "privileged"
+    if category not in valid_categories:
+        return {"success": False, "message": _(f"Invalid category. Must be one of: {valid_categories}")}
+
+    try:
+        # Validate tool exists
+        plugin_manager = get_plugin_manager()
+        tool_registry = get_tool_registry()
+        all_tools = plugin_manager.get_all_tools()
+
+        # Include external tools from hooks
+        external_tools = tool_registry._get_external_tools()
+        all_tools.update(external_tools)
+
+        if tool_name not in all_tools:
+            return {"success": False, "message": _(f"Tool '{tool_name}' not found")}
+
+        # Get or create tool configuration
+        if frappe.db.exists("FAC Tool Configuration", tool_name):
+            config = frappe.get_doc("FAC Tool Configuration", tool_name)
+        else:
+            # Create new configuration
+            tool_info = all_tools[tool_name]
+            auto_category = detect_tool_category(tool_info.instance)
+
+            config = frappe.new_doc("FAC Tool Configuration")
+            config.tool_name = tool_name
+            config.plugin_name = tool_info.plugin_name
+            config.description = tool_info.description
+            config.enabled = 1
+            config.auto_detected_category = auto_category
+            config.source_app = getattr(tool_info.instance, "source_app", "frappe_assistant_core")
+
+        config.tool_category = category
+        config.category_override = 1 if override else 0
+        config.save(ignore_permissions=True)
+
+        frappe.db.commit()
+
+        # Clear caches
+        tool_registry.clear_cache()
+
+        return {"success": True, "message": _(f"Tool '{tool_name}' category updated to '{category}'")}
+
+    except Exception as e:
+        frappe.log_error(f"Failed to update tool category for '{tool_name}': {str(e)}")
+        return {"success": False, "message": _(f"Error: {str(e)}")}
+
+
+@frappe.whitelist(methods=["POST"])
+def update_tool_role_access(tool_name: str, role_access_mode: str, roles: list = None):
+    """
+    Update role access settings for a tool.
+
+    Args:
+        tool_name: The name of the tool
+        role_access_mode: "Allow All" or "Restrict to Listed Roles"
+        roles: List of dicts with {role, allow_access} for restricted mode
+
+    Returns:
+        Success status and message
+    """
+    frappe.only_for(["System Manager", "Assistant Admin"])
+    from frappe_assistant_core.core.tool_registry import get_tool_registry
+    from frappe_assistant_core.utils.plugin_manager import get_plugin_manager
+    from frappe_assistant_core.utils.tool_category_detector import detect_tool_category
+
+    valid_modes = ["Allow All", "Restrict to Listed Roles"]
+    if role_access_mode not in valid_modes:
+        return {"success": False, "message": _(f"Invalid mode. Must be one of: {valid_modes}")}
+
+    if isinstance(roles, str):
+        import json
+
+        roles = json.loads(roles)
+
+    try:
+        # Validate tool exists
+        plugin_manager = get_plugin_manager()
+        tool_registry = get_tool_registry()
+        all_tools = plugin_manager.get_all_tools()
+
+        # Include external tools from hooks
+        external_tools = tool_registry._get_external_tools()
+        all_tools.update(external_tools)
+
+        if tool_name not in all_tools:
+            return {"success": False, "message": _(f"Tool '{tool_name}' not found")}
+
+        # Get or create tool configuration
+        if frappe.db.exists("FAC Tool Configuration", tool_name):
+            config = frappe.get_doc("FAC Tool Configuration", tool_name)
+        else:
+            # Create new configuration
+            tool_info = all_tools[tool_name]
+            category = detect_tool_category(tool_info.instance)
+
+            config = frappe.new_doc("FAC Tool Configuration")
+            config.tool_name = tool_name
+            config.plugin_name = tool_info.plugin_name
+            config.description = tool_info.description
+            config.enabled = 1
+            config.tool_category = category
+            config.auto_detected_category = category
+            config.source_app = getattr(tool_info.instance, "source_app", "frappe_assistant_core")
+
+        config.role_access_mode = role_access_mode
+
+        # Update role access table if in restricted mode
+        if role_access_mode == "Restrict to Listed Roles" and roles:
+            # Clear existing role access
+            config.role_access = []
+
+            # Add new role access entries
+            for role_entry in roles:
+                if isinstance(role_entry, dict):
+                    config.append(
+                        "role_access",
+                        {"role": role_entry.get("role"), "allow_access": role_entry.get("allow_access", 1)},
+                    )
+
+        config.save(ignore_permissions=True)
+        frappe.db.commit()
+
+        # Clear caches
+        tool_registry.clear_cache()
+
+        return {"success": True, "message": _(f"Tool '{tool_name}' role access updated")}
+
+    except Exception as e:
+        frappe.log_error(f"Failed to update tool role access for '{tool_name}': {str(e)}")
+        return {"success": False, "message": _(f"Error: {str(e)}")}
+
+
+@frappe.whitelist()
+def get_available_roles() -> dict:
+    """
+    Get list of available roles for role access configuration.
+
+    Returns:
+        List of roles with names
+    """
+    frappe.only_for(["System Manager", "Assistant Admin"])
+    try:
+        roles = frappe.get_all(
+            "Role",
+            filters={"disabled": 0, "desk_access": 1},
+            fields=["name", "role_name"],
+            order_by="name",
+        )
+
+        # Also include some system roles that might be useful
+        system_roles = ["System Manager", "Administrator"]
+        for sr in system_roles:
+            if not any(r["name"] == sr for r in roles):
+                roles.append({"name": sr, "role_name": sr})
+
+        return {"success": True, "roles": roles}
+
+    except Exception as e:
+        frappe.log_error(f"Failed to get available roles: {str(e)}")
+        return {"success": False, "error": str(e), "roles": []}
